fix: move .typelib and .gir dirs into private dir
Currently the file `Xviewer-3.0.typelib` is located in `$libdir/girepository-1.0`,
i.e. `/usr/lib/x86_64-linux-gnu/girepository-1.0/`.

Therefore, it only works as long as the install $prefix is `/usr` but fails for e.g. `/usr/local`
with `Error loading Xviewer typelib: Typelib file for namespace 'Xviewer', version '3.0' not found`
as this path is not in the default g_ir search path.

This commit changes the path to the xviewer pkglibdir, e.g. `/usr/local/lib/x86_64-linux-gnu/xviewer/girepository-1.0/`
and changes the code to load it from there.
This behavior is similar to Xed's.

Also, the path of the .gir file is moved into the private data dir for consistency.