Make color correction optional and disabled by default
It's not OK for the viewer to show corrected colors based on profiles
which are not always selected by the user.

Color profiles are autogenerated based on EDID, leading to xviewer
showing different colors than what's defined in the file.

This is really bad for anyone working on screenshots, or color codes.

This option is an opt-in in Gimp. It doesn't exist in Pix or in most
other applications.

Fixes https://github.com/linuxmint/xviewer/issues/80
https://askubuntu.com/questions/1236599/images-are-too-dark-and-saturated-in-image-viewer
https://gitlab.gnome.org/GNOME/eog/-/issues/289
https://bugzilla.gnome.org/show_bug.cgi?id=675645#c11