Revert 7a7b55b (add mutex locks...)
This reverts:
ce1505dab7
7a7b55bbe1

As these changes can cause deadlocks on the main UI thread, by
the UI thread locking the store, then attempting to access it
in a later callback from the same thread, among other issues.

The original code used a mutex to lock changes made by the thumb-
nailer, but not for widget access, which is sensible. I've been
unable to reproduce the original issue this patch was meant to
fix (though I don't doubt there are access issues, as I've observed
assertion errors caused by attempting to re-thumbnail already
-loaded or already-thumbnailing images). Unfortunately the new
issues are worse than the original, at least from an end-user
standpoint.

For now, revert the patch, until time can be spent on the problem.